[FEATURE] More implementation/prototype
Have a running prototype using a manually programmed game in the game
builder. Changed dt from a long to a double (since it should be a
ratio). Implemented a clock. Changed some arrays to lists for
simplicity.  Made a dummy main method and a dummy output window. Added
ChangeVelAction to change velocity in one/both direction components.
Fixed a few bugs in the BorderEvent event and in the GameObject step
method.