[BUGFIX] Collision Fixes
Fixed many of the issues with colliding a circle with a rectangle
(demonstrating this in the levels.xml by having the ball reflect off the
rectangle paddles).

Fixed the issue of the ball passing through the player controlled
bumper. This was caused by the input event being evaluated last, when
really it should be the first thing to occur. A number of other
safe-guard were made to handle edge cases that cause things to display
weirdly. This introduces some new funny behaviour (such as the bumper
and the ball fighting for dominance and not moving until the player
releases the input), but it's better than before I think.

Couldn't regularly replicate the problem of the ball being stuck in a
corner, and thus wasn't able to fix it.

Also fixed a crash that occurred by having an event without any actions.